# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Caffe-Backend",
        "version": "0.1"
    },
    "paths": {
...
        "/api/mesas": {
            "get": {
                "tags": [
                    "Mesas"
                ],
                "summary": "Listar mesas",
                "operationId": "5bd7c941a94d9c293b50cbf1877cbf4d",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página (opcional)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna a lista de mesas"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Mesas"
                ],
                "summary": "Adicionar mesa",
                "operationId": "53d4743d0c3ba89ee2f5ecb95b775927",
                "requestBody": {
                    "description": "Informações da nova mesa",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Status_mesa",
                                    "Preco_total",
                                    "Numero_mesa",
                                    "Tipo_pagamento"
                                ],
                                "properties": {
                                    "Status_mesa": {
                                        "type": "boolean"
                                    },
                                    "Preco_total": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "Numero_mesa": {
                                        "type": "integer"
                                    },
                                    "Tipo_pagamento": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mesa adicionada com sucesso"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/api/mesas/{id}": {
            "get": {
                "tags": [
                    "Mesas"
                ],
                "summary": "Mostrar detalhes da mesa",
                "operationId": "e1a4805b4bc8fdc565a149d7d91d78e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da mesa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna os detalhes da mesa"
                    },
                    "404": {
                        "description": "Mesa não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Mesas"
                ],
                "summary": "Atualizar mesa",
                "operationId": "b1fc5e1d7e4dcd89d09cb24235b3cf07",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da mesa a ser atualizada",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Informações da mesa a serem atualizadas",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Status_mesa": {
                                        "type": "boolean"
                                    },
                                    "Preco_total": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "Numero_mesa": {
                                        "type": "integer"
                                    },
                                    "Tipo_pagamento": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mesa atualizada com sucesso"
                    },
                    "404": {
                        "description": "Mesa não encontrada"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/api/mesas/{id}/abrir-fechar": {
            "put": {
                "tags": [
                    "Mesas"
                ],
                "summary": "Abrir ou fechar mesa",
                "operationId": "95e400f7f1ffc196ce954c851b4ddeed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da mesa a ser aberta ou fechada",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Status da mesa",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Status_mesa": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status da mesa atualizado com sucesso"
                    },
                    "404": {
                        "description": "Mesa não encontrada"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
...
    }
}